+++ Firestore/core/src/firebase/firestore/local/leveldb_migrations.cc
@@ -22,6 +22,7 @@
 #include "Firestore/Protos/nanopb/firestore/local/mutation.nanopb.h"
 #include "Firestore/Protos/nanopb/firestore/local/target.nanopb.h"
 #include "Firestore/core/src/firebase/firestore/local/leveldb_key.h"
+#include "Firestore/core/src/firebase/firestore/local/memory_index_manager.h"
 #include "Firestore/core/src/firebase/firestore/model/document_key.h"
 #include "Firestore/core/src/firebase/firestore/model/types.h"
 #include "Firestore/core/src/firebase/firestore/nanopb/reader.h"
@@ -36,6 +37,8 @@ using leveldb::Iterator;
 using leveldb::Slice;
 using leveldb::Status;
 using leveldb::WriteOptions;
+using model::DocumentKey;
+using model::ResourcePath;
 using nanopb::Reader;
 using nanopb::Writer;
 
@@ -60,8 +63,9 @@ namespace {
  *   * Migration 4 ensures that every document in the remote document cache
  *     has a sentinel row with a sequence number.
  *   * Migration 5 drops held write acks.
+ *   * Migration 6 populates the collection_parents index.
  */
-const LevelDbMigrations::SchemaVersion kSchemaVersion = 5;
+const LevelDbMigrations::SchemaVersion kSchemaVersion = 6;
 
 /**
  * Save the given version number as the current version of the schema of the
@@ -214,6 +218,8 @@ void EnsureSentinelRow(LevelDbTransaction* transaction,
 }
 
 /**
+ * Migration 4.
+ *
  * Ensure each document in the remote document table has a corresponding
  * sentinel row in the document target index.
  */
@@ -238,6 +244,65 @@ void EnsureSentinelRows(leveldb::DB* db) {
                       sentinel_value);
   }
   SaveVersion(4, &transaction);
+  transaction.Commit();
+}
+
+// Helper to add an index entry iff we haven't already written it (as determined
+// by the provided cache).
+void EnsureCollectionParentRow(LevelDbTransaction* transaction,
+                               MemoryCollectionParentIndex* cache,
+                               const DocumentKey& key) {
+  const ResourcePath& collection_path = key.path().PopLast();
+  if (cache->Add(collection_path)) {
+    std::string collection_id = collection_path.last_segment();
+    ResourcePath parent_path = collection_path.PopLast();
+
+    std::string key =
+        LevelDbCollectionParentKey::Key(collection_id, parent_path);
+    std::string empty_buffer;
+    transaction->Put(key, empty_buffer);
+  }
+}
+
+/**
+ * Migration 6.
+ *
+ * Creates appropriate LevelDbCollectionParentKey rows for all collections
+ * of documents in the remote document cache and mutation queue.
+ */
+void EnsureCollectionParentsIndex(leveldb::DB* db) {
+  LevelDbTransaction transaction(db, "Ensure Collection Parents Index");
+
+  MemoryCollectionParentIndex cache;
+
+  // Index existing remote documents.
+  std::string documents_prefix = LevelDbRemoteDocumentKey::KeyPrefix();
+  auto it = transaction.NewIterator();
+  it->Seek(documents_prefix);
+  LevelDbRemoteDocumentKey document_key;
+  for (; it->Valid() && absl::StartsWith(it->key(), documents_prefix);
+       it->Next()) {
+    HARD_ASSERT(document_key.Decode(it->key()),
+                "Failed to decode document key");
+
+    EnsureCollectionParentRow(&transaction, &cache,
+                              document_key.document_key());
+  }
+
+  // Index existing mutations.
+  std::string mutations_prefix = LevelDbDocumentMutationKey::KeyPrefix();
+  it = transaction.NewIterator();
+  it->Seek(mutations_prefix);
+  LevelDbDocumentMutationKey key;
+  for (; it->Valid() && absl::StartsWith(it->key(), mutations_prefix);
+       it->Next()) {
+    HARD_ASSERT(key.Decode(it->key()),
+                "Failed to decode document-mutation key");
+
+    EnsureCollectionParentRow(&transaction, &cache, key.document_key());
+  }
+
+  SaveVersion(6, &transaction);
   transaction.Commit();
 }
 
@@ -286,6 +351,10 @@ void LevelDbMigrations::RunMigrations(leveldb::DB* db,
 
   if (from_version < 5 && to_version >= 5) {
     RemoveAcknowledgedMutations(db);
+  }
+
+  if (from_version < 6 && to_version >= 6) {
+    EnsureCollectionParentsIndex(db);
   }
 }